Refactor: update flow defaultCatch extension usage
