+++ features/register/onboard/src/main/kotlin/team/applemango/runnerbe/feature/register/onboard/component/step/TermsTable.kt
@@ -47,20 +47,24 @@ import team.applemango.runnerbe.feature.register.onboard.asset.StringAsset
 import team.applemango.runnerbe.feature.register.onboard.util.Web
 import team.applemango.runnerbe.shared.android.constant.DataStoreKey
 import team.applemango.runnerbe.shared.android.extension.dataStore
-import team.applemango.runnerbe.shared.android.extension.defaultCatch
 import team.applemango.runnerbe.shared.compose.default.RunnerbeCheckBoxDefaults
-import team.applemango.runnerbe.shared.compose.extension.activityViewModel
 import team.applemango.runnerbe.shared.compose.extension.presentationDrawableOf
 import team.applemango.runnerbe.shared.compose.theme.ColorAsset
 import team.applemango.runnerbe.shared.compose.theme.Typography
+import team.applemango.runnerbe.shared.domain.extension.defaultCatch
 
 private val VerticalPadding = 25.dp
 private val HorizontalPadding = 12.dp
 private val TermsTableShape = RoundedCornerShape(10.dp)
 
+// `vm: OnboardViewModel = activityViewModels()` 안한 이유:
+// DFM 는 의존성이 반대로 되서 hilt 를 사용하지 못함
+// 따라서 직접 factory 로 인자를 주입해 줘야 함
+// 이는 OnboardActivity 에서 해주고 있으므로,
+// OnboardActivity 에서 vm 를 가져와야 함
 @Composable
 internal fun TermsTable(
-    vm: OnboardViewModel = activityViewModel(),
+    vm: OnboardViewModel,
     onAllTermsCheckStateChanged: (allChecked: Boolean) -> Unit,
 ) {
     val context = LocalContext.current.applicationContext
@@ -103,7 +107,7 @@ internal fun TermsTable(
         val preferences = context
             .dataStore
             .data
-            .defaultCatch(vm = vm)
+            .defaultCatch(action = vm::emitException)
             .first()
 
         if (preferences[DataStoreKey.Onboard.TermsAllCheck] == true) {
