+++ shared/android/src/main/kotlin/team/applemango/runnerbe/shared/android/extension/flow.kt
@@ -15,8 +15,6 @@ import androidx.lifecycle.flowWithLifecycle
 import androidx.lifecycle.lifecycleScope
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.catch
-import team.applemango.runnerbe.shared.android.base.BaseViewModel
 
 fun <T> Flow<T>.collectWithLifecycle(
     lifecycleOwner: LifecycleOwner,
@@ -31,13 +29,4 @@ fun <T> Flow<T>.collectWithLifecycle(
             action(value)
         }
     }
-}
-
-fun <T> Flow<T>.defaultCatch(
-    name: String? = null,
-    vm: BaseViewModel
-) = catch { cause ->
-    val exception =
-        IllegalStateException("${name ?: cause.tag} flow collect exception: ${cause.message}")
-    vm.emitException(exception)
 }
